Allow the `docker-proxy` to be overridden
This patch adds support for a new command-line argument
`--userland-proxy-bin=""` alongside the existing `--userland-proxy`
argument.

A replacement proxy binary is a convenient place to perform some
side-effect such as opening a tunnel to expose the port in some
other place over some other protocol. For example where docker is
running in a VM on a desktop, a replacement docker-proxy can use some
out-of-band mechanism to bind a port on the host, and send/receive
the data over some host<->VM channel such as AF_VSOCK.

Other (less reliable) ways I've tried to solve the same problem include:

- watch docker API events and create/destroy tunnels on demand.
  Unfortunately errors are handled asynchronously and cannot cause
  the container start to fail.

- proxy the docker API itself and create/destroy tunnels during the
  API calls. This has the downside that it won't create tunnels when
  containers are automatically started by the daemon, and the
  docker API is larger (and faster moving) than the `docker-proxy` "API".

The new --userland-proxy-bin argument is only inspected if
--userland-proxy=true.

If --userland-proxy-bin="" (the default) then the existing built-in
docker-proxy will be used.

Signed-off-by: David Scott <dave.scott@docker.com>